third push. bot. added /history, /favorites, updated README.md

+++ handlers/message_handlers.py
@@ -1,3 +1,7 @@
+import re
+
+import emoji
+
 from utils.misc.class_User import UserProfile
 from utils.misc.class_FSM import UserStates
 from utils import cities_names
@@ -6,6 +10,7 @@ from aiogram.dispatcher.storage import FSMContext
 from aiogram.types import Message
 from keyboards.reply import simple_keyboard
 from keyboards.inline import inline_keyboard
+from database.base_inition.users_db import get_user_history
 
 
 @dp.message_handler(commands=['start'])
@@ -17,12 +22,12 @@ async def command_start(message: Message) -> None:
     await message.delete()
     await message.answer('Привет! Я бот по поиску отелей.'
                          '\n\nДоступные команды для начала работы:\n'
-                         '/start - Команда приветствия\n'
+                         '/start - Команда приветствия\n\n'
                          '/lowprice - поиск отелей в указанном городе (сортировка от низкой цены)\n'
                          '/highprice - поиск отелей в указанном городе (сортировка от высокой цены)\n'
                          '/bestdeal - поиск отелей в заданном диапазоне цен, с мин-ным расстоянием от центра города\n'
-                         '/history - показать историю (команда, дата/время, результаты поиска)\n'
-                         '\nУдачи!',
+                         '/history - показать историю (команда, дата/время, результаты поиска)\n\n'
+                         '/cancel - выйти из режима поиска\n\nУдачи!',
                          reply_markup=simple_keyboard.show_commands_keyboard())
 
 
@@ -38,6 +43,38 @@ async def command_cancel(message: Message, state: FSMContext) -> None:
         return
     else:
         await state.finish()
+
+
+@dp.message_handler(commands=['favorites'])
+async def show_history(message: Message) -> None:
+    if message.from_user.id not in UserProfile.all_users:
+        new_user = UserProfile(message.from_user.id, '0', '/start')
+    cur_user = UserProfile.all_users[message.from_user.id]
+
+    some_resp = get_user_history('favorites', message.from_user.id)
+    for i_elem in some_resp:
+        if i_elem != 'NOT OK':
+            hotel_name = re.findall(r'\b(.+)\n', i_elem)[0]
+            await bot.send_message(message.from_user.id, i_elem,
+                                   reply_markup=inline_keyboard.hide_last_message(hotel_name=hotel_name[5::]))
+        else:
+            await bot.send_message(message.from_user.id, 'Похоже, что история еще не заполнена.\n'
+                                                         'Попробуйте что-то поискать.')
+
+
+@dp.message_handler(commands=['history'])
+async def show_history(message: Message) -> None:
+    if message.from_user.id not in UserProfile.all_users:
+        new_user = UserProfile(message.from_user.id, '0', '/start')
+    cur_user = UserProfile.all_users[message.from_user.id]
+
+    some_resp = get_user_history('history', message.from_user.id)
+    for i_elem in some_resp:
+        if i_elem != 'NOT OK':
+            await bot.send_message(message.from_user.id, i_elem, reply_markup=inline_keyboard.hide_last_message())
+        else:
+            await bot.send_message(message.from_user.id, 'Похоже, что история еще не заполнена.\n'
+                                                         'Попробуйте что-то поискать.')
 
 
 @dp.message_handler(commands=['lowprice', 'highprice', 'bestdeal'])
@@ -92,17 +129,25 @@ async def get_hotels_quantity(message: Message, state: FSMContext) -> None:
     # если цифры, то сохраняем количество отелей, устанавливаем arr_date_year состояние
     cur_user = UserProfile.all_users[message.from_user.id]
     await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
-    if message.text == '0' or int(message.text) < 0:
+    if int(message.text) <= 0:
         cur_user.message_to_delete = await message.answer('Меня не проведешь. Ноль или отрицательное - нельзя.\n '
                                                           'Давайте ещё раз:')
         await message.delete()
-    cur_user.cities_quantity = message.text
-    cur_user.message_to_delete = await bot.send_message(message.from_user.id,
-                                                        'Необходимо выбрать дату заезда.\nПожалуйста, укажите год:',
-                                                        reply_markup=inline_keyboard.year_calendar_keyboard()
-                                                        )
-    await message.delete()
-    await UserStates.next()
+    else:
+        if cur_user.status[1] == '/bestdeal':
+            cur_user.message_to_delete = await bot.send_message(message.from_user.id,
+                                                                emoji.emojize('Укажите минимальную :dollar banknote: '
+                                                                              'цену за ночь:'))
+            await UserStates.min_night_price.set()
+        else:
+            cur_user.cities_quantity = message.text
+            cur_user.message_to_delete = await bot.send_message(message.from_user.id,
+                                                                emoji.emojize('Необходимо выбрать дату заезда.\n'
+                                                                              'Пожалуйста, укажите год: :calendar:'),
+                                                                reply_markup=inline_keyboard.year_calendar_keyboard()
+                                                                )
+            await message.delete()
+            await UserStates.next()
 
 
 @dp.message_handler(lambda message: not message.text.isdigit(), state=UserStates.min_night_price)
@@ -118,17 +163,19 @@ async def check_set_min_price(message: Message, state: FSMContext) -> None:
 async def check_set_min_price(message: Message, state: FSMContext) -> None:
     # если режим /bestdeal, ловим мин цену за ночь
     cur_user = UserProfile.all_users[message.from_user.id]
-    if int(message.text) < 0:
+    if int(message.text) <= 0:
         await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
         await message.delete()
         cur_user.message_to_delete = await message.answer('Меня не обманешь.\n'
-                                                          'Пожалуйста, используйте только не отрицательные цифры.')
+                                                          'Пожалуйста, используйте только не отрицательные цифры.\n'
+                                                          'Ноль тоже нельзя.')
     else:
         cur_user.low = message.text
         await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
         await message.delete()
-        cur_user.message_to_delete = await message.answer('Пожалуйста, укажите максимальную цену за ночь:')
-        await UserStates.next()
+        cur_user.message_to_delete = await message.answer(emoji.emojize('Пожалуйста, укажите максимальную '
+                                                                        ':dollar banknote: цену за ночь:'))
+        await UserStates.max_night_price.set()
 
 
 @dp.message_handler(lambda message: not message.text.isdigit(), state=UserStates.max_night_price)
@@ -144,7 +191,7 @@ async def check_set_min_price(message: Message, state: FSMContext) -> None:
 async def check_set_min_price(message: Message, state: FSMContext) -> None:
     # если режим /bestdeal, ловим макс цену за ночь, если она меньше low, меняем местами
     cur_user = UserProfile.all_users[message.from_user.id]
-    if int(message.text) < 0:
+    if int(message.text) <= 0:
         await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
         await message.delete()
         cur_user.message_to_delete = await message.answer('Меня не обманешь.\n'
@@ -154,21 +201,23 @@ async def check_set_min_price(message: Message, state: FSMContext) -> None:
             cur_user.low, cur_user.high = message.text, cur_user.low
             await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
             await message.delete()
-            await UserStates.show_results.set()
-            cur_user.message_to_delete = await bot.send_message(message.from_user.id,
+            cur_user.message_to_delete = await bot.send_message(message.from_user.id, emoji.emojize(
                                                                 'Немного перепутали min и max. Я поменяю.\n'
-                                                                'Параметры сохранены.\nНажмите "показать".',
-                                                                reply_markup=inline_keyboard.show_result()
+                                                                'Теперь выберите дату заезда. :calendar:\n'
+                                                                'Для начала год:'),
+                                                                reply_markup=inline_keyboard.year_calendar_keyboard()
                                                                 )
+            await UserStates.arr_date_year.set()
         else:
             cur_user.high = message.text
             await bot.delete_message(chat_id=message.from_user.id, message_id=cur_user.message_to_delete.message_id)
             await message.delete()
-            await UserStates.show_results.set()
-            cur_user.message_to_delete = await bot.send_message(message.from_user.id,
-                                                                'Параметры сохранены.\nНажмите "показать".',
-                                                                reply_markup=inline_keyboard.show_result()
+            cur_user.message_to_delete = await bot.send_message(message.from_user.id, emoji.emojize(
+                                                                'Теперь выберите дату заезда. :calendar:\n'
+                                                                'Для начала год:'),
+                                                                reply_markup=inline_keyboard.year_calendar_keyboard()
                                                                 )
+            await UserStates.arr_date_year.set()
 
 
 @dp.message_handler(content_types='text')
